update gabor funcs
